% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChrColumn}
\alias{getChrColumn}
\title{getChrColumn}
\usage{
getChrColumn(df, col=NULL, needed=TRUE)
}
\arguments{
\item{df}{(data.frame or equivalent) The data frame were columns are searched}

\item{col}{(number of character) If a number, it will be returned. If a character, it will be treated as a the exact name of the column. If a column with that name exists, its position in the data.frame will be returned. If it does not, NULL will be returned and an error might be raised (depending on the value of \code{needed}). If NULL, the column will be searched using a predefined pattern (defaults to NULL)}

\item{needed}{Whether the column is needed or not. If TRUE, an error will be raised if the column is not found. (defaults to TRUE)}
}
\value{
The number of the column matching the specification or NULL if no column was found.
}
\description{
Identify the column in a data.frame with the chromosome information
}
\details{
Identify the column of a data.frame that contains the chromosome 
information and return its position
}
\examples{

getChrColumn(df)
getChrColumn(df, "strange.name")

}
