% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getColumn}
\alias{getColumn}
\title{getColumn}
\usage{
getColumn(df, col=NULL, pattern="",  msg.col.name="", needed=TRUE)
}
\arguments{
\item{df}{(data.frame or equivalent) The data frame were columns are searched}

\item{col}{(number of character) If a number, it will be returned. If a character, it will be treated as a the exact name of the column. If a column with that name exists, its position in the data.frame will be returned. If it does not, NULL will be returned and an error might be raised (depending on the value of \code{needed}). If NULL, teh column will be searched using pattern (defaults to NULL)}

\item{pattern}{(character) The pattern to match the column name. If more than one column matches the pattern, the first one (leftmost) will be returned. The pattern may be any valid regular expression. (defaults to "")}

\item{msg.col.name}{(character) Only used in the error message to make the message clearer. The name of the column we are searching for. (defaults to "")}

\item{needed}{(defaults to TRUE)}
}
\value{
The number of the column matching the specification or NULL if no column was found.
}
\description{
Use simple pattern matching to try to identify a column in a data.frame by 
its name
}
\details{
This function will use pattern matching to try to identify which column of 
a data.frame contains a certain data. It will return its number. If 
\code{col} is specified it will not use pattern matching but identify
if a column with that exact name exists and return its position.
}
\examples{

df <- data.frame("id"= "rs1234","chromosome"="chr1", "Start"=0, "end.position"=100,
"copy.number.level"=3, "LOH"=0, "median.value.per.segment"=1.2,
"BAF"=0.2, "Log Ratio"=1.5, "strange.name"="strange.value")

getColumn(df, pattern="Chr|chr",  msg.col.name="Chromosome", needed=TRUE)
getColumn(df, col="chromosome",  msg.col.name="Chromosome", needed=TRUE)

}
