% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCopyNumberColors.R
\name{getCopyNumberColors}
\alias{getCopyNumberColors}
\title{getCopyNumberColors}
\usage{
getCopyNumberColors()
}
\arguments{
\item{colors}{}
}
\value{
A vector with colors
}
\description{
Get the colors representing each copy number
}
\details{

}
\examples{



}
