% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCopyNumberCalls.cnmops.R
\name{loadCopyNumberCalls.cnmops}
\alias{loadCopyNumberCalls.cnmops}
\title{loadCopyNumberCalls.cnmops}
\usage{
loadCopyNumberCalls.cnmops(cn.mops.res, chr.col = NULL, start.col = NULL, end.col = NULL, cn.col = "CN", segment.value.col = "median", genome = NULL, verbose = TRUE)
}
\arguments{
\item{cn.mops.res}{The name of the file with the data or the name of the variable with the data}

\item{chr.col}{The name or number of the column with chromosome information. If NULL, it is automatically identified. (default to NULL)}

\item{start.col}{The name or number of the column with start position information. If NULL, it is automatically identified. (default to NULL)}

\item{end.col}{The name or number of the column with end position information. If NULL, it is automatically identified. (default to NULL)}

\item{cn.col}{The name or number of the column with CN information. If NULL, it is automatically identified. (default to "CN")}

\item{segment.value.col}{The name or number of the column with segment value. It can be mean or median. If NULL, it is automatically identified. (default to "median")}

\item{genome}{The name of the genome (default to NULL)}

\item{verbose}{Wether information messages should be generated. (default to TRUE)}
}
\value{
A GRanges object with a range per copy number segment
}
\description{
Loads copy number calls from either cn.mops result class or in a tabular format
}
\details{
This function will load segments data from either cn.mops result class or
in a tabular format or file. Internally it uses the \code{toGRanges} function from regioneR
package and can work with any format accepted by it, including R objects
and local or remote files. 
Column names of the format loaded are specified as default but,
if no column names are specified, it will use simple heuristics to try to
identify the relevant data columns.
}
\note{
The returned GRanges will have the chromosome names following the UCSC style
irrespective of the original format.
}
\examples{
## loadCopyNumberCalls.cnmops from cn.mops result class:
data(cn.mops, package = "cn.mops")

cn.mops.res <- cn.mops(XRanges)
cn.mops.res <- calcIntegerCopyNumbers(cn.mops.res)
loadCopyNumberCalls.cnmops(cn.mops.res = cn.mops.res)  

## loadCopyNumberCalls.cnmops from a file where cn.mops result was saved:
cn.mops.res <- "cn.mops.segmentation.csv"
loadCopyNumberCalls.cnmops(cn.mops.res = cn.mops.res) 

}
