% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCopyNumberCalls.DECoN.R
\name{loadCopyNumberCalls.DECoN}
\alias{loadCopyNumberCalls.DECoN}
\title{loadCopyNumberCalls.DECoN}
\usage{
loadCopyNumberCalls.DECoN(decon.file, chr.col = NULL, start.col = "Start", end.col = "End", cn.col = NA, segment.value.col = NULL, genome = NULL, verbose = TRUE)
}
\arguments{
\item{decon.file}{The name of the file with the data}

\item{chr.col}{The name or number of the column with chromosome information. If NULL, it is automatically identified. (default to NULL)}

\item{start.col}{The name or number of the column with start position information. If NULL, it is automatically identified. (default to "Start")}

\item{end.col}{The name or number of the column with end position information. If NULL, it is automatically identified. (default to "End")}

\item{cn.col}{The name or number of the column with CN information. If NULL, it is automatically identified. (default to NA)}

\item{segment.value.col}{The name or number of the column with segment value. If NULL, it is automatically identified. (default to NULL)}

\item{genome}{The name of the genome (default to NULL)}

\item{verbose}{Wether information messages should be generated. (default to TRUE)}
}
\value{
A GRanges object with a range per copy number segment
}
\description{
Loads copy number calls from DECoN  output file
}
\details{
This function will load segments data from DECoN output  file. 
Internally it uses the \code{toGRanges} function from regioneR
package and can work with any format accepted by it, including R objects
and local or remote files. 
If no column names are specified, it will use simple heuristics to try to
identify the relevant data columns.
}
\examples{
decon.file <- system.file("extdata", "DECoN_output.txt", package = "CopyNumberPlots", mustWork = TRUE)
loadCopyNumberCalls.DECoN(decon.file = decon.file)

}
