% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCopyNumberCalls.pennCNV.R
\name{loadCopyNumberCalls.pennCNV}
\alias{loadCopyNumberCalls.pennCNV}
\title{loadCopyNumberCalls.seg}
\usage{
loadCopyNumberCalls.pennCNV(pennCNV.file, chr.col = NULL, start.col = NULL, end.col = NULL, segment.value.col = NA, genome = NULL, verbose = TRUE)
}
\arguments{
\item{pennCNV.file}{The name of the file with the data}

\item{chr.col}{The name or number of the column with chromosome information. If NULL, it is automatically identified. (default to NULL)}

\item{start.col}{The name or number of the column with start position information. If NULL, it is automatically identified. (default to NULL)}

\item{end.col}{The name or number of the column with end position information. If NULL, it is automatically identified. (default to NULL)}

\item{cn.col}{The name or number of the column with CN information. If NULL, it is automatically identified. (default to NA)}

\item{segment.value.col}{The name or number of the column with segment value. If NULL, it is automatically identified. (default to NA)}

\item{genome}{The name of the genome (default to NULL)}

\item{verbose}{Wether information messages should be generated. (default to TRUE)}
}
\value{
A GRanges object with a range per copy number segment
}
\description{
Loads copy number calls from pennCNV.rawcnv file format
}
\details{
This function will load copy number calls from pennCNV.rawcnv file format. 
Internally it uses the \code{toGRanges} function from regioneR
package and can work with any format accepted by it, including R objects
and local or remote files. 
If no column names are specified, it will use simple heuristics to try to
identify the relevant data columns.
}
\note{
The returned GRanges will have the chromosome names following the UCSC style
irrespective of the original format.
}
\examples{
## loadCopyNumberCalls.pennCNV from .rawcnv file format: The file to run the example can be found in: http://penncnv.openbioinformatics.org/en/latest/user-guide/test/
## under example.seg file name.
pennfile <- "testFiles/pennCNV.rawcnv"
loadCopyNumberCalls.pennCNV(pennCNV.file = pennCNV.file)

}
