% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCopyNumberCalls.R
\name{loadCopyNumberCalls}
\alias{loadCopyNumberCalls}
\title{loadCopyNumberCalls}
\usage{
loadCopyNumberCalls(cnv.data, chr.col=NULL, start.col=NULL, end.col=NULL, cn.col=NULL, loh.col=NULL, segment.value.col=NULL, genome=NULL, verbose=TRUE)
}
\arguments{
\item{cnv.data}{The name of the file with the data}

\item{chr.col}{The name or number of the column with chromosome information. If NULL, it is automatically identified. (default to NULL)}

\item{start.col}{The name or number of the column with start position information. If NULL, it is automatically identified. (default to NULL)}

\item{end.col}{The name or number of the column with end position information. If NULL, it is automatically identified. (default to NULL)}

\item{cn.col}{The name or number of the column with BAF information. If NULL, it is automatically identified. (default to NULL)}

\item{loh.col}{The name or number of the column with LRR information. If NULL, it is automatically identified. (default to NULL)}

\item{segment.value.col}{The name or number of the column with SNP identifier information. If NULL, it is automatically identified. (default to NULL)}

\item{genome}{The name of the genome (default to NULL)}

\item{verbose}{Wether information messages should be generated. (default to TRUE)}
}
\value{
A GRanges object with a range per copy number segment
}
\description{
Loads copy number calls from a tabular format
}
\details{
This function will load segments data from any "bed-like" data structure
in R or file. Internally it uses the \code{toGRanges} function from regioneR
package and can work with any format accepted by it, including R objects
and local or remote files.
If no column names are specified, it will use simple heuristics to try to
identify the relevant data columns.
}
\note{
The returned GRanges will have the chromosome names following the UCSC style
irrespective of the original format.
}
\examples{



}
