% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCopyNumberCalls.seg.R
\name{loadCopyNumberCalls.seg}
\alias{loadCopyNumberCalls.seg}
\title{loadCopyNumberCalls.seg}
\usage{
loadCopyNumberCalls.seg (seg.file, chr.col = "chrom", start.col = "loc.start", end.col = "loc.end", segment.value.col = "seg.mean", genome = NULL, chr.transformation = "23:X,24:Y,25:MT", verbose = TRUE)
}
\arguments{
\item{seg.file}{The name of the file with the data}

\item{chr.col}{The name or number of the column with chromosome information. If NULL, it is automatically identified. (default to "chrom")}

\item{start.col}{The name or number of the column with start position information. If NULL, it is automatically identified. (default to "loc.start")}

\item{end.col}{The name or number of the column with end position information. If NULL, it is automatically identified. (default to "loc.end")}

\item{segment.value.col}{The name or number of the column with segment value. If NULL, it is automatically identified. (default to "seg.mean")}

\item{genome}{The name of the genome (default to NULL)}

\item{chr.transformation}{Transformation of the chromosomes. The files.seg have the chromosomes in a numerical form.
In order to transform the sexual and mitocondrial chromosomes from numbers to characters,
the user has to provide in the chr.transformation parameter as describe in 
https://cnvkit.readthedocs.io/en/stable/importexport.html#import-seg, in a python dictionary format ("key:value").(default to "23:X,24:Y,25:MT" )}

\item{verbose}{Wether information messages should be generated. (default to TRUE)}
}
\value{
A GRanges object with a range per copy number segment
}
\description{
Loads copy number calls from .seg file format
}
\details{
This function will load segments data from .seg file format. 
Internally it uses the \code{toGRanges} function from regioneR
package and can work with any format accepted by it, including R objects
and local or remote files. 
If no column names are specified, it will use simple heuristics to try to
identify the relevant data columns.
}
\examples{
## loadCopyNumberCalls.seg from .seg file format: 
## the file to run the example can be found in: https://software.broadinstitute.org/software/igv/SEG
## under example.seg file name.

seg.file <- system.file("extdata", "DNACopy_output.seg", package = "CopyNumberPlots", mustWork = TRUE)
loadCopyNumberCalls.seg(seg.file)

}
