% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSNPDataFromVCF.R
\name{loadSNPDataFromVCF}
\alias{loadSNPDataFromVCF}
\title{loadSNPDataFromVCF}
\usage{
loadSNPDataFromVCF(vcf.file, regions=NULL, genome="hg19", mirror.baf=TRUE, verbose=TRUE)
}
\arguments{
\item{vcf.file}{The name of the file with the data}

\item{regions}{The regions to which we will limit the import (defaults to NULL)}

\item{genome}{(a character)The name of the genome (defaults to "hg19")}

\item{mirror.baf}{Flip the baf of about half the snps (the ones in odd positions in the genome) to achieve a mirror-like effect as in SNP arrays (defaults to TRUE)}

\item{verbose}{Wether information messages should be generated. (defaults to TRUE)}
}
\value{
A GRanges object with a range per SNP
}
\description{
Loads SNP array data in a tabular format
}
\details{
Given a file name, the function loads SNP array data in a tabular format.
It will try to identify the columns with the relevant information
(chr, position, BAF, LRR, etc...) or will use the column number or name
supplied by the user, if any. It will convert the tabular data into a
GRanges, with one range per SNP in the table.
}
\examples{

vcf.file <- system.file("extdata", "example.vcf.gz", package = "CopyNumberPlots", mustWork = TRUE)
snps <- loadSNPDataFromVCF(vcf.file)

kp <- plotKaryotype(plot.type = 4)
plotBAF(kp, snps = snps, labels = names(snps))

}
