% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCopyNumberCalls.R
\name{plotCopyNumberCalls}
\alias{plotCopyNumberCalls}
\title{plotCopyNumberCalls}
\usage{
plotCopyNumberCalls()
}
\arguments{
\item{karyoplot}{}
}
\value{
Invisibly returns the karyoplot object representing the plot. With it
it is possible to add other elements to the plot using standrad karyoploteR
functions
}
\description{
Plot the segments representing the copy number calls by any algorithm
}
\details{
Plots the segments
}
\examples{



}
