% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCopyNumberCallsAsLines.R
\name{plotCopyNumberCallsAsLines}
\alias{plotCopyNumberCallsAsLines}
\title{plotCopyNumberCallsAsLines}
\usage{
plotCopyNumberCalls()
}
\arguments{
\item{karyoplot}{}
}
\value{
Invisibly returns the karyoplot object representing the plot. With it
it is possible to add other elements to the plot using standrad karyoploteR
functions
}
\description{
Plot the segments representing the copy number calls by any algorithm
}
\details{
Plots the segments
}
\examples{

set.seed(1234)
gg <- filterChromosomes(getGenome("hg19"))

scnas <- createRandomRegions(40, 10e6, 10e6, non.overlapping=TRUE)
scnas$cn <- floor(runif(40, 0, 4))
scnas$loh <- ifelse(scnas$cn<2, 1, 0)
normal.regs <- subtractRegions(gg, scnas)
normal.regs$cn <- 2
normal.regs$loh <- 0
scnas <- sort(c(scnas, normal.regs))

kp <- plotKaryotype("hg19")
plotCopyNumberCallsAsLines(kp, scnas)

kp <- plotKaryotype("hg19", chromosomes="chr3")
plotCopyNumberCalls(kp, scnas, r0=0, r1=0.2)
plotCopyNumberCallsAsLines(kp, scnas, r0=0.25, r1=0.55)
plotCopyNumberCallsAsLines(kp, scnas, r0=0.65, r1=0.95, style="segments", col="red", ymin=0, ymax=4, numticks=3)

kp <- plotKaryotype("hg19")
plotCopyNumberCallsAsLines(kp, scnas, r0=0.2, r1=0.4, style="segments", axis=FALSE, label="CopyNumber")

}
