% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLRR.R
\name{plotLRR}
\alias{plotLRR}
\title{plotLRR}
\usage{
plotSNPData(snps, main="Raw Data", chromosomes="canonical", zoom=NULL, lrr.min=-4, lrr.max=2, total.height=1, bottom=0, margin=0.05, points.cex=0.3, labels.cex=1.5, main.cex=2, axis.cex=1.2, chr.cex=1.5)
}
\arguments{
\item{snps}{The SNP array data}

\item{points.cex}{(defaults to 0.3)}

\item{axis.cex}{(defaults to 1.2)}

\item{main}{(defaults to "Raw Data")}

\item{chromosomes}{(defaults to "canonical")}

\item{zoom}{(defaults to NULL)}

\item{lrr.min}{(defaults to -4)}

\item{lrr.max}{(defaults to 2)}

\item{total.height}{(defaults to 1)}

\item{bottom}{(defaults to 0)}

\item{margin}{(defaults to 0.05)}

\item{labels.cex}{(defaults to 1.5)}

\item{main.cex}{(defaults to 2)}

\item{chr.cex}{(defaults to 1.5)}
}
\value{
Invisibly returns the karyoplot object representing the plot. With it
it is possible to add other elements to the plot using standrad karyoploteR
functions
}
\description{
Plots the raw SNP array data using karyoploteR
}
\details{
Creates a plot with the LRR values along the genome
}
\examples{

pos <- floor(runif(1000, 1, 10000))
lrr.data <- toGRanges("chr1", pos, pos)
lrr.data$lrr <- rnorm(1000, mean = 0, sd = 0.3)

kp <- plotKaryotype(zoom=toGRanges("chr1", 1, 10000))
plotLRR(kp, lrr.data)


names(mcols(lrr.data)) <- "values"
kp <- plotKaryotype(zoom=toGRanges("chr1", 1, 10000))
plotLRR(kp, lrr.data, lrr.column="values", r0=0, r1=0.5, points.col="red", label="Values")
plotLRR(kp, lrr.data, lrr.column=1, r0=0.5, r1=1, points.col="orange", label="First", label.cex=0.8, points.cex=1.4)


}
