% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readHDF5Ztree}
\alias{readHDF5Ztree}
\title{readHDF5Ztree}
\usage{
readHDF5Ztree(hdf5.file)
}
\arguments{
\item{hdf5.file}{(H5IdComponent object) The result of opening an HDF5 file 
with rhdf5::H5Fopen(data.file). The file must contain a "tree" element 
with Ztree. For example, HDF5 files for CNV data produced by 10X CellRanger.}
}
\value{
A valid hclust object (standard R hierarchical clustering results)
}
\description{
Read the Ztree (scipy matrix representarion of a hierarchical clustering)
contained in a 10X single-cell CNV HDF5 file and return a
valid R hclust object representing the same tree.

More info on the Ztree matrix: \url{https://joernhees.de/blog/2015/08/26/scipy-hierarchical-clustering-and-dendrogram-tutorial/}
}
\examples{

#Not Run
data.file <- "path to a single-cell CNV data file"
open.file <- rhdf5::H5Fopen(data.file)
hc.tree <- readHDF5Ztree(open.file)

}
