% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformChr}
\alias{transformChr}
\title{transformChr}
\usage{
transformChr(seg.data, chr.transformation = "23:X,24:Y,25:MT")
}
\arguments{
\item{seg.data}{variable containing a data.frame with the information}

\item{chr.transformation}{parameter where the transformation of the chromosomes is established
using a python dictionary format ("key:value").(default to "23:X,24:Y,25:MT")}
}
\value{
A data.frame object where the name of the chromosomes will have been changed according to chr.transformation parameter.
}
\description{
Transformation of the chromosomes
}
\details{
This function will transform the name of the chormosomes, according to the parameter chr.transformation. 
There are files, such as files.seg, which have the chromosomes in a number format.
In order to transform the sexual and mitocondrial chromosomes from numbers to characters,
the user has to provide in the chr.transformation parameter as describe in 
https://cnvkit.readthedocs.io/en/stable/importexport.html#import-seg, in a python dictionary format ("key:value")
}
\examples{
seg.file <- system.file("extdata", "DNACopy_output.seg", package = "CopyNumberPlots", mustWork = TRUE)
seg.data <- read.table(file = seg.file, sep = "\\t", skip = 1, stringsAsFactors = FALSE)
colnames(seg.data) <-  c("ID", "chrom", "loc.start", "loc.end", "num.mark", "seg.mean")

# here we have the name of the chromosomes in a number format and how many segments there are in each one.
table(seg.data$chrom)

# by this way we have the name 23 and 24 respectively transformed by X and Y. 
table(transformChr(seg.data = seg.data, chr.transformation ="23:X,24:Y,25:MT"))

}
