% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSNPDataFromVCF.R
\name{loadSNPDataFromVCF}
\alias{loadSNPDataFromVCF}
\title{loadSNPDataFromVCF}
\usage{
loadSNPDataFromVCF(vcf.file, genome="hg19", randomize.baf=TRUE, verbose=TRUE)
}
\arguments{
\item{vcf.file}{The name of the file with the data}

\item{genome}{The name of the genome (default to "hg19")}

\item{randomize.baf}{Flip the baf of about half the snps (the ones in odd positions in the genome) to achieve a mirror-like effect as in SNP arrays (defaults to TRUE)}

\item{verbose}{Wether information messages should be generated. (default to TRUE)}
}
\value{
A GRanges object with a range per SNP
}
\description{
Loads SNP array data in a tabular format
}
\details{
Given a file name, the function loads SNP array data in a tabular format.
It will try to identify the columns with the relevant information
(chr, position, BAF, LRR, etc...) or will use the column number or name
supplied by the user, if any. It will convert the tabular data into a
GRanges, with one range per SNP in the table.
}
\note{
The returned GRanges will have the chromosome names following the UCSC style
irrespective of the original format.
}
\examples{



}
