% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCopyNumberCalls.R
\name{plotCopyNumberCalls}
\alias{plotCopyNumberCalls}
\title{plotCopyNumberCalls}
\usage{
plotCopyNumberCalls()
}
\arguments{
\item{karyoplot}{}
}
\value{
Invisibly returns the karyoplot object representing the plot. With it
it is possible to add other elements to the plot using standrad karyoploteR
functions
}
\description{
Plot the segments representing the copy number calls by any algorithm
}
\details{
Plots the segments
}
\examples{


gg <- filterChromosomes(getGenome("hg19"))

scnas <- createRandomRegions(40, 10e6, 10e6)
scnas$cn <- floor(runif(40, 0, 4))
normal.regs <- subtractRegions(gg, scnas)
normal.regs$cn <- 2
scnas <- sort(c(scnas, normal.regs))
scnas$loh <- ifelse(scnas$cn<2, TRUE, FALSE)

kp <- plotKaryotype("hg19")
plotCopyNumberCalls(kp, scnas)

kp <- plotKaryotype("hg19")
plotCopyNumberCalls(kp, scnas, cn.colors="red_blue")


}
