% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLRR.R
\name{plotLRR}
\alias{plotLRR}
\title{plotLRR}
\usage{
plotLRR(karyoplot, snps, lrr.column="lrr", label="LRR", ymin=-4, ymax=2, out.of.range = "points", out.of.range.col="red", density.height=0.05, density.window=1e5,
               line.at.0 = TRUE, line.at.0.col="blue", r0=0, r1=1, points.cex=0.3, points.col="#333333",
               points.pch=16, label.cex=1.5, label.srt=90, label.margin=0.03, axis.cex=1.2, ...)
}
\arguments{
\item{karyoplot}{the karyoplot}

\item{snps}{the data}

\item{lrr.column}{(defaults to "lrr")}

\item{label}{(defaults to "LRR")}

\item{ymin}{(defaults to -4)}

\item{ymax}{(defaults to 2)}

\item{out.of.range}{(defaults to  "points")}

\item{out.of.range.col}{(defaults to "red")}

\item{density.height}{(defaults to 0.05)}

\item{density.window}{(defaults to 1e5)}

\item{line.at.0}{(defaults to  TRUE)}

\item{line.at.0.col}{(defaults to "blue")}

\item{r0}{(defaults to 0)}

\item{r1}{(defaults to 1)}

\item{points.cex}{(defaults to 0.3)}

\item{points.col}{(defaults to "#333333")}

\item{points.pch}{(defaults to 16)}

\item{label.cex}{(defaults to 1.5)}

\item{label.srt}{(defaults to 90)}

\item{label.margin}{(defaults to 0.03)}

\item{axis.cex}{(defaults to 1.2)}

\item{...}{Additional params}
}
\value{
Invisibly returns the karyoplot object representing the plot. With it
it is possible to add other elements to the plot using standrad karyoploteR
functions
}
\description{
Plots the raw SNP array data using karyoploteR
}
\details{
Creates a plot with the LRR values along the genome
}
\examples{

pos <- floor(runif(1000, 1, 10000))
lrr.data <- toGRanges("chr1", pos, pos)
lrr.data$lrr <- rnorm(1000, mean = 0, sd = 0.3)

kp <- plotKaryotype(zoom=toGRanges("chr1", 1, 10000))
plotLRR(kp, lrr.data)


names(mcols(lrr.data)) <- "values"
kp <- plotKaryotype(zoom=toGRanges("chr1", 1, 10000))
plotLRR(kp, lrr.data, lrr.column="values", r0=0, r1=0.5, points.col="red", label="Values")
plotLRR(kp, lrr.data, lrr.column=1, r0=0.5, r1=1, points.col="orange", label="First", label.cex=0.8, points.cex=1.4)


}
